cloudappclient: handle invalid protocol for skill (#337)
