+++ test/apps/cloudAppClient/manager.test.js
